requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
git+https://github.com/openai/CLIP.git
